django_backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pymongo==4.6.3
dnspython==2.6.1
python-dotenv==1.0.1
whitenoise==6.7.0
django-cors-headers==4.4.0
Pillow==10.4.0
openpyxl==3.1.5
bcrypt==4.1.2